Extract log in and sign up
