+++ src/components/signupForm/SignUpForm.module.css
@@ -33,10 +33,6 @@ form > * {
 
 .toggle {
   border-color: var(--secondary-color-site);
-  width: 100%;
-  padding: 12px 10px;
-  margin: 8px 0;
-  font-weight: bold;
 }
 
 /* Potentially whack button hover flair */
